Fully remove abstractMacros
